requirements.txt
arrow==1.4.0
asgiref==3.10.0
binaryornot==0.4.4
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
cookiecutter==2.6.0
cryptography==46.0.3
Django==5.2.7
django-admin-sortable2==2.2.8
django-tailwind==4.2.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
pycparser==2.23
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.0.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
